# carwash-backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Car Wash Admin API Documentation",
        "description": "API documentation for Car Wash Admin backend application (Laravel/Sanctum)",
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/v1/services": {
            "get": {
                "tags": [
                    "Services"
                ],
                "summary": "Get list of all services",
                "operationId": "67cbbb0f9c56d57d64701620f46d07d6",
                "parameters": [
                    {
                        "name": "filter[is_active]",
                        "in": "query",
                        "description": "Filter by active status",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of services",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ServiceResource"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Services"
                ],
                "summary": "Create a new service",
                "operationId": "35e7c832ed23b7ce6e0b4c2ed5b7adef",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "price"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Yangi yuvish"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": "35000.00"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Yangi xizmat tavsifi"
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Service created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ServiceResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/v1/services/{service}": {
            "put": {
                "tags": [
                    "Services"
                ],
                "summary": "Update an existing service",
                "operationId": "2a81679d59168ecad190579fbcde506d",
                "parameters": [
                    {
                        "name": "service",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Yangilangan yuvish"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": "40000.00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Service updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ServiceResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Services"
                ],
                "summary": "Delete a specific service",
                "operationId": "95c11607891043b37d98de7e517847ff",
                "parameters": [
                    {
                        "name": "service",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Service deleted successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/v1/shifts/open": {
            "post": {
                "tags": [
                    "Shifts"
                ],
                "summary": "Open a new shift for the authenticated user (Kassirning yangi smenasini ochish)",
                "operationId": "9433507cc6b16b260d5ee7552124bde7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "opening_cash"
                                ],
                                "properties": {
                                    "opening_cash": {
                                        "description": "Smena boshlanishidagi kassa qoldig'i",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Shift opened successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ShiftResource"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "User already has an open shift",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "You already have an open shift."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized action"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/v1/shifts/close": {
            "put": {
                "tags": [
                    "Shifts"
                ],
                "summary": "Close the currently open shift for the authenticated user (Ochiq smenani yopish)",
                "operationId": "92bd96fa29e079f83fa683e70704d8de",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "closing_cash"
                                ],
                                "properties": {
                                    "closing_cash": {
                                        "description": "Smena yopilishidagi kassa qoldig'i",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Shift closed successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ShiftResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No open shift found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No query results for model [App\\Models\\Shift."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized action"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ServiceResource": {
                "title": "Service Resource",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Standart yuvish"
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": "25000.00"
                    },
                    "description": {
                        "type": "string",
                        "example": "Avtomobil tashqi qismini tozalash"
                    },
                    "isActive": {
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
...
            "ShiftResource": {
                "title": "Shift Resource",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "userId": {
                        "type": "integer",
                        "example": 2
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "open",
                            "closed"
                        ],
                        "example": "open"
                    },
                    "openTime": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "closeTime": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter Sanctum token for authentication (Bearer Token)",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}